backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
google-generativeai==0.8.3
python-dotenv==1.0.0
httpx==0.26.0
openai==1.54.0

